Deepfake-Detection-Plugin-Chrome-Extension-main/requirements.txt
flask
flask-cors
torch
torchvision
facenet-pytorch
transformers
pillow
opencv-python
numpy